refact (controllers): rename namespace and controllers
