--- GestaoFinanceira/GestaoFinanceira/Controller/CrtSubCategories.cs
@@ -1,30 +1,0 @@
-﻿using GestaoFinanceira.Model;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GestaoFinanceira.Controller
-{
-    class CrtSubCategories
-    {
-        public List<SubCategories> List()
-        {
-            throw new NotImplementedException();
-        }
-        public SubCategories Find(int categoriesId)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void Save(SubCategories subCategories)
-        {
-            throw new NotImplementedException();
-        }
-        public void Remove(SubCategories subCategories)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
