+++ GestaoFinanceira/GestaoFinanceira/Controllers/CategoriesController.cs
@@ -1,0 +1,31 @@
+﻿using GestaoFinanceira.Model;
+using System;
+using System.CodeDom;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GestaoFinanceira.Controllers
+{
+    class CategoriesController
+    {
+        public List<Categories> List()
+        {
+            throw new NotImplementedException();
+        }
+        public Categories Find(int categoriesId)
+        {
+            throw new NotImplementedException();
+        }
+
+        public void Save(Categories categories)
+        {
+            throw new NotImplementedException();
+        }
+        public void Remove(Categories categories)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
