+++ GestaoFinanceira/GestaoFinanceira/Controllers/SubCategoriesController.cs
@@ -1,0 +1,30 @@
+﻿using GestaoFinanceira.Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GestaoFinanceira.Controllers
+{
+    class CrtSubCategories
+    {
+        public List<SubCategories> List()
+        {
+            throw new NotImplementedException();
+        }
+        public SubCategories Find(int categoriesId)
+        {
+            throw new NotImplementedException();
+        }
+
+        public void Save(SubCategories subCategories)
+        {
+            throw new NotImplementedException();
+        }
+        public void Remove(SubCategories subCategories)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
